Use `NotificationHelper` in media scanner too
